printer/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# printer/print.py: 5
Pillow == 8.0.1

# printer/print.py: 7
# printer/print_tester.py: 4
beautifulsoup4 == 4.9.3

# printer/print.py: 4
purepng == 0.2.0

# printer/print.py: 4
pypng == 0.0.20

# printer/print.py: 3
# printer/print_tester.py: 3
pyqrcode == 1.2.1

# printer/print.py: 8
# printer/print_tester.py: 5
requests == 2.25.0

# printer/print.py: 6
thermalprinter == 0.2.0
